feat: subscriptions management bot
